fix(cache): preserve packed DWARF and compiler flag order (#1418, #1419)
