+++ crates/zccache-depgraph/src/rustc_args/README.md
@@ -1,0 +1,5 @@
+# Rust compiler argument parser tests
+
+This directory contains focused unit tests for the Rust compiler argument
+parser in the parent module. It is split out to keep production source below
+the repository line-count ceiling.
